Fix/authorize user

+++ app/controllers/searches_controller.rb
@@ -23,14 +23,18 @@ class SearchesController < ApplicationController
 
   def show
     @search = Search.find(params[:id])
-    @areas = @search.areas
-    @results = @search.results.unfavorited.includes(:listing)
-    @favorited_results = @search.results.favorited.includes(:listing)
+    if user_is_owner?(@search)
+      @areas = @search.areas
+      @results = @search.results.unfavorited.includes(:listing)
+      @favorited_results = @search.results.favorited.includes(:listing)
+    else
+      redirect_to dashboard_path
+    end
   end
 
   def destroy
     search = Search.find(params[:id])
-    search.destroy!
+    search.destroy! if user_is_owner?(search)
     redirect_to dashboard_path
   end
 
@@ -49,5 +53,9 @@ class SearchesController < ApplicationController
                                    :max_price,
                                    :refresh_interval,
                                    area_ids: []
+  end
+
+  def user_is_owner?(search)
+    current_user.id == search.user_id
   end
 end
